+++ App.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { StyleSheet, Text, View, } from 'react-native';
+
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>This is your first screen. Change this text. 😁</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  }
+});
