--- tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "jsx": "react",
-    "lib": [
-      "ES6"
-    ],
-    "moduleResolution": "node",
-    "noEmit": true,
-    "target": "esnext",
-    "noImplicitAny": false
-  },
-  "exclude": [
-    "node_modules",
-    "babel.config.js",
-    "metro.config.js",
-    "jest.config.js",
-    "__tests__/**/*-test.ts"
-  ],
-  "extends": "expo/tsconfig.base"
-}
